main_project/backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.0
redis==5.0.0
aioredis==2.0.1
sqlalchemy==2.0.23
pymysql==1.1.0
google-cloud-pubsub==2.18.0
google-cloud-storage==2.10.0
google-cloud-bigquery==3.11.0
pyspark==3.5.0
google-auth==2.23.4
pytest==7.4.2
httpx==0.25.1
requests==2.31.0
loguru==0.7.0
gunicorn==21.2.0
cryptography==41.0.7